requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
django-oauth2-provider==0.2.6.1
djangorestframework==3.12.2
django-cors-headers==3.7.0
django-oauth-toolkit==2.2.0
django-filter==22.1
Pillow==9.4.0 
gunicorn==20.1.0